backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
kiteconnect==5.0.1
apscheduler==3.10.4
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
